+++ pillarbox-demo/src/main/java/ch/srgssr/pillarbox/demo/ui/showcases/ShowcasesHome.kt
@@ -77,6 +77,18 @@ fun ShowcasesHome(navController: NavController) {
                 modifier = itemModifier,
                 onClick = { navController.navigate(NavigationRoutes.story) }
             )
+
+            HorizontalDivider()
+
+            DemoListItemView(
+                title = stringResource(R.string.chapters),
+                modifier = itemModifier,
+                onClick = {
+                    navController.navigate(
+                        NavigationRoutes.chapters
+                    )
+                }
+            )
         }
 
         DemoListHeaderView(
@@ -198,17 +210,6 @@ fun ShowcasesHome(navController: NavController) {
                 onClick = {
                     navController.navigate(
                         NavigationRoutes.video360
-                    )
-                }
-            )
-            HorizontalDivider()
-
-            DemoListItemView(
-                title = stringResource(R.string.chapters),
-                modifier = itemModifier,
-                onClick = {
-                    navController.navigate(
-                        NavigationRoutes.chapters
                     )
                 }
             )
